requirements.txt
chromadb>=0.5.0
openai>=1.50.0
anthropic>=0.30.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
markdown>=3.6
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
requests>=2.32.0
numpy>=1.26.0
torch>=2.4.0
transformers>=4.40.0
requests
